templates/requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
requests==2.31.0
python-socketio==5.10.0
gunicorn==21.2.0
eventlet==0.33.3
